Update configs to be able to hold multiple experiments to effectively group all the experiments one might be interested in.
