+++ repsim/configs/hierarchical_vision_shortcuts.yaml
@@ -1,0 +1,24 @@
+# Holds all information needed to run experiments.
+measures: all
+experiments:
+  - name: Vision Short Cut Experiments
+    type: GroupSeparationExperiment  # Defines the type of experiment class to use
+    representation_dataset: "ColorDot_0_CIFAR10DataModule"
+    filter_key_vals:  # Only keeps models that have any of the values for the keys.
+      train_dataset:
+        - ColorDot_100_CIFAR10DataModule
+        - ColorDot_75_CIFAR10DataModule
+        - ColorDot_50_CIFAR10DataModule
+        - ColorDot_25_CIFAR10DataModule
+        - ColorDot_0_CIFAR10DataModule
+      seed:
+        - 0
+        - 1
+        - 2
+        - 3
+        - 4
+      domain: VISION
+    grouping_keys:  # Separates the trained models by this key into groups for the `GroupSeparationExperiment`.
+      - train_dataset
+    separation_keys:  # Creates a separate experiment for each unique value of the key.
+      - architecture
