+++ repsim/run.py
@@ -5,13 +5,27 @@ from typing import get_args
 
 import yaml
 from loguru import logger
-from repsim.benchmark.archive.shortcut_experiment import GroupSeparationExperiment
-from repsim.benchmark.model_selection import get_grouped_trained_models
+from repsim.benchmark.abstract_experiment import AbstractExperiment
+from repsim.benchmark.group_separation_experiment import GroupSeparationExperiment
+from repsim.benchmark.model_selection import get_grouped_models
+from repsim.benchmark.monotonicity_experiment import MonotonicityExperiment
+from repsim.benchmark.multimodel_experiments import MultiModelExperiment
 from repsim.benchmark.registry import ALL_TRAINED_MODELS
 from repsim.benchmark.registry import DOMAIN_TYPE
 from repsim.benchmark.registry import TrainedModel
-from repsim.measures.measure_index import ALL_MEASURES
-from tqdm import tqdm
+from repsim.measures import ALL_MEASURES
+from repsim.measures.utils import SimilarityMeasure
+
+
+def get_experiment_from_name(name: str) -> AbstractExperiment:
+    if name == "GroupSeparationExperiment":
+        return GroupSeparationExperiment
+    elif name == "MonotonicityExperiment":
+        return MonotonicityExperiment
+    elif name == "MultiModelExperiment":
+        return MultiModelExperiment
+    else:
+        raise ValueError(f"Invalid experiment name: {name}")
 
 
 def read_yaml_config(config_path: str) -> dict:
@@ -22,60 +36,77 @@ def read_yaml_config(config_path: str) -> dict:
         config = yaml.safe_load(file)
 
     # ------------ If no measures provided we default to all measures ------------ #
-    if "measures" not in config:
-        logger.info("No measures provided in config. Defaulting to all measures")
-        config["measures"] = list(ALL_MEASURES.keys())
     return config
+
+
+def get_measures(config: dict) -> list[SimilarityMeasure]:
+    if config["measures"] == "all":
+        return list(ALL_MEASURES.values())
+    else:
+        return [ALL_MEASURES[measurename] for measurename in config["measures"]]
 
 
 def verify_config(config: dict) -> None:
     """
     Raise error if config is not valid
     """
-    assert config["domain"] in get_args(DOMAIN_TYPE), "Invalid domain, choose from: {}".format(DOMAIN_TYPE)
-    # assert config["experiment"] in EXPERIMENT_IDENTIFIER, "Invalid experiment, choose from: {}".format(
-    #     EXPERIMENT_IDENTIFIER
-    # )
-    assert all(
-        [measurename in ALL_MEASURES for measurename in config["measures"]]
-    ), f"Invalid measures. Got {[f'{c}' for c in config['measures']]} choose from: {list(ALL_MEASURES.keys())}"
-    assert "representation_dataset" in config, "No representation dataset provided"
+    assert "measures" in config, "No measures provided"
+    assert isinstance(config["measures"], (str, list[str])), "Measures should be a string or a list of strings"
+    if isinstance(config["measures"], str):
+        measure_name = config["measures"]
+        assert measure_name in ["all"] + list(ALL_MEASURES.keys()), f"Invalid measure: {measure_name}"
+    else:
+        measure_name = config["measures"]
+        assert all([m in ALL_MEASURES for m in measure_name]), f"Invalid measures provided: {measure_name}"
 
-    filter_key_vals = config.get("filter_key_vals", None)
-    if filter_key_vals:
-        for key, val in filter_key_vals.items():
-            assert isinstance(key, str), "Key should be a string"
-            assert isinstance(val, str) or isinstance(val, list), "Value should be a string or a list of strings"
+    assert "experiments" in config, "No experiments in config."
+    # for exp in config["experiments"]:
+    #     filter_key_vals = config.get("filter_key_vals", None)
+    #     if filter_key_vals:
+    #         for key, val in filter_key_vals.items():
+    #             assert isinstance(key, str), "Key should be a string"
+    #             assert isinstance(val, str) or isinstance(val, list), "Value should be a string or a list of strings"
 
-    differentiation_keys = config.get("differentiation_keys", None)
-    if differentiation_keys:
-        for key in differentiation_keys:
-            assert isinstance(key, str), "Differentiation key should be a string"
+    #     differentiation_keys = config.get("differentiation_keys", None)
+    #     if differentiation_keys:
+    #         for key in differentiation_keys:
+    #             assert isinstance(key, str), "Differentiation key should be a string"
 
 
 def run(config_path: str):
     config = read_yaml_config(config_path)
     verify_config(config)
     logger.debug("Config is valid")
-    filter_key_vals = config.get("filter_key_vals", None)
-    differentiation_keys = config.get("differentiation_keys", None)
-    # Not quite sure how to best make this work with the entire setup.
-    #   Some form of hierarchy would need to be added, but that doesn't exist atm.
-    #   Maybe create a hierarchical nested structure and make this a recursive function? to auto-aggregate?
-    grouped_models: list[Sequence[TrainedModel]] = get_grouped_trained_models(
-        ALL_TRAINED_MODELS,
-        filter_key_vals,
-        differentiation_keys,
-    )
-    measures = [ALL_MEASURES[measurename] for measurename in config["measures"]]
+    measures = get_measures(config)
+
+    all_experiments = []
+    for experiment in config["experiments"]:
+        if experiment["type"] == "GroupSeparationExperiment":
+            filter_key_vals = experiment.get("filter_key_vals", None)
+            grouping_keys = experiment.get("grouping_keys", None)
+            separation_keys = experiment.get("separation_keys", None)
+
+            # Not quite sure how to best make this work with the entire setup.
+            #   Some form of hierarchy would need to be added, but that doesn't exist atm.
+            #   Maybe create a hierarchical nested structure and make this a recursive function? to auto-aggregate?
+            grouped_models: list[list[Sequence[TrainedModel]]] = get_grouped_models(
+                models=ALL_TRAINED_MODELS,
+                filter_key_vals=filter_key_vals,
+                separation_keys=separation_keys,
+                grouping_keys=grouping_keys,
+            )
+            for group in grouped_models:
+                exp = GroupSeparationExperiment(
+                    grouped_models=group,
+                    measures=measures,
+                    representation_dataset=experiment["representation_dataset"],
+                )
+                all_experiments.append(exp)
+
     # -------------------- Now compare/eval the grouped models ------------------- #
-    exp = GroupSeparationExperiment(
-        grouped_models=grouped_models,
-        measures=measures,  # We should likely introduce a "TAG" system to select measures
-        representation_dataset=config["representation_dataset"],
-    )
-    exp.run()
-    exp.eval()
+    for ex in all_experiments:
+        ex.run()
+        ex.eval()
 
 
 if __name__ == "__main__":
@@ -87,5 +118,5 @@ if __name__ == "__main__":
     #     )
     #     args = parser.parse_args()
     #     logger.debug("Parsing config"
-    config_path = os.path.join(os.path.dirname(__file__), "configs", "vision_shortcuts.yaml")
+    config_path = os.path.join(os.path.dirname(__file__), "configs", "hierarchical_vision_shortcuts.yaml")
     run(config_path)
